+++ server/include/bussiness.h
@@ -6,14 +6,12 @@
 #include "log.h"
 #include "mysqloperate.h"
 #include "task.h"
+#include "transfer.h"
 
 typedef struct {
     int length;
     char data[4096];
 } DataBlock;
-
-int sendn(int sockfd, void* buf, int length);
-int recvn(int sockfd, void* buf, int length);
 
 int sendFile(int sockfd, int fd, off_t f_size);
 int recvFile(int sockfd, MYSQL* mysql, int uid);
@@ -27,10 +25,6 @@ void pwdCmd(Task* task);
 int getsCmd(Task* task);
 int putsCmd(Task* task);
 void mkdirCmd(Task* task);
-void regCheck1(Task* task);
-void regCheck2(Task* task);
-void loginCheck1(Task* task);
-void loginCheck2(Task* task);
 void unknownCmd(void);
 
 int taskHandler(Task* task);
