+++ server/src/task.c
@@ -1,6 +1,6 @@
 #include "../include/task.h"
 
-void taskFree(Task* task) {
+void freeTask(Task* task) {
     for (int i = 0; task->args[i] != NULL; ++i) {
         free(task->args[i]);
     }
